test: split shared/explorer.test in twain
